upd script and initialize contracts
